Update/1.14.0

+++ materialious/src/lib/api/index.ts
@@ -1,4 +1,4 @@
-import { getVideoTYjs } from '$lib/api/youtubejs';
+import { getVideoYTjs } from '$lib/api/youtubejs/video';
 import { Capacitor } from '@capacitor/core';
 import { get } from 'svelte/store';
 import {
@@ -13,15 +13,10 @@ import {
 	synciousInstanceStore
 } from '../store';
 import type {
-	Channel,
-	ChannelContentPlaylists,
-	ChannelContentVideos,
 	ChannelPage,
 	Comments,
 	DeArrow,
 	Feed,
-	HashTag,
-	Playlist,
 	PlaylistPage,
 	ResolvedUrl,
 	ReturnYTDislikes,
@@ -29,8 +24,28 @@ import type {
 	Subscription,
 	ApiExntendedProgressModel,
 	Video,
-	VideoPlay
+	VideoPlay,
+	SearchOptions,
+	SearchResults,
+	CommentsOptions,
+	ChannelOptions,
+	ChannelContent
 } from './model';
+import { commentsSetDefaults, searchSetDefaults, useEngineFallback } from './misc';
+import { getSearchYTjs } from './youtubejs/search';
+import { isYTBackend } from '$lib/misc';
+import { getSearchSuggestionsYTjs } from './youtubejs/searchSuggestions';
+import { getResolveUrlYTjs } from './youtubejs/misc';
+import { getCommentsYTjs } from './youtubejs/comments';
+import { getChannelContentYTjs, getChannelYTjs } from './youtubejs/channel';
+import {
+	amSubscribedYTjs,
+	deleteUnsubscribeYTjs,
+	getFeedYTjs,
+	getSubscriptionsYTjs,
+	postSubscribeYTjs
+} from './youtubejs/subscriptions';
+import { getPlaylistYTjs } from './youtubejs/playlist';
 
 export function buildPath(path: string): URL {
 	return new URL(`${get(instanceStore)}/api/v1/${path}`);
@@ -71,11 +86,20 @@ export function buildAuthHeaders(): { headers: Record<string, string> } {
 }
 
 export async function getPopular(fetchOptions?: RequestInit): Promise<Video[]> {
+	// Doesn't exist in YTjs.
+	if (isYTBackend()) {
+		return [];
+	}
+
 	const resp = await fetchErrorHandle(await fetch(buildPath('popular'), fetchOptions));
 	return await resp.json();
 }
 
 export async function getResolveUrl(url: string): Promise<ResolvedUrl> {
+	if (isYTBackend() || useEngineFallback('ResolveUrl')) {
+		return await getResolveUrlYTjs(url);
+	}
+
 	const resp = await fetchErrorHandle(await fetch(`${buildPath('resolveurl')}?url=${url}`));
 	return await resp.json();
 }
@@ -85,14 +109,18 @@ export async function getVideo(
 	local: boolean = false,
 	fetchOptions?: RequestInit
 ): Promise<VideoPlay> {
-	if (get(playerYouTubeJsAlways) && Capacitor.isNativePlatform()) {
-		return await getVideoTYjs(videoId);
+	if (
+		(get(playerYouTubeJsAlways) && Capacitor.isNativePlatform()) ||
+		isYTBackend() ||
+		useEngineFallback('Video')
+	) {
+		return await getVideoYTjs(videoId);
 	}
 
 	const resp = await fetch(setRegion(buildPath(`videos/${videoId}?local=${local}`)), fetchOptions);
 
 	if (!resp.ok && get(playerYouTubeJsFallback) && Capacitor.isNativePlatform()) {
-		return await getVideoTYjs(videoId);
+		return await getVideoYTjs(videoId);
 	} else {
 		await fetchErrorHandle(resp);
 	}
@@ -111,23 +139,17 @@ export async function getDislikes(
 
 export async function getComments(
 	videoId: string,
-	parameters: {
-		sort_by?: 'top' | 'new';
-		source?: 'youtube' | 'reddit';
-		continuation?: string;
-	},
+	options: CommentsOptions,
 	fetchOptions?: RequestInit
 ): Promise<Comments> {
-	if (typeof parameters.sort_by === 'undefined') {
-		parameters.sort_by = 'top';
-	}
+	commentsSetDefaults(options);
 
-	if (typeof parameters.source === 'undefined') {
-		parameters.source = 'youtube';
+	if (isYTBackend() || useEngineFallback('Comments')) {
+		return await getCommentsYTjs(videoId, options);
 	}
 
 	const path = buildPath(`comments/${videoId}`);
-	path.search = new URLSearchParams(parameters).toString();
+	path.search = new URLSearchParams(options).toString();
 	const resp = await fetchErrorHandle(await fetch(path, fetchOptions));
 	return await resp.json();
 }
@@ -136,32 +158,32 @@ export async function getChannel(
 	channelId: string,
 	fetchOptions?: RequestInit
 ): Promise<ChannelPage> {
+	if (isYTBackend() || useEngineFallback('Channel')) {
+		return getChannelYTjs(channelId);
+	}
 	const resp = await fetchErrorHandle(
 		await fetch(buildPath(`channels/${channelId}`), fetchOptions)
 	);
 	return await resp.json();
 }
 
-export type channelSortBy = 'oldest' | 'newest' | 'popular';
-export type channelContentTypes = 'videos' | 'playlists' | 'streams' | 'shorts';
-
 export async function getChannelContent(
 	channelId: string,
-	parameters: {
-		type?: channelContentTypes;
-		continuation?: string;
-		sortBy?: channelSortBy;
-	},
+	options: ChannelOptions,
 	fetchOptions?: RequestInit
-): Promise<ChannelContentVideos | ChannelContentPlaylists> {
-	if (typeof parameters.type === 'undefined') parameters.type = 'videos';
+): Promise<ChannelContent> {
+	if (typeof options.type === 'undefined') options.type = 'videos';
 
-	const url = buildPath(`channels/${channelId}/${parameters.type}`);
+	const url = buildPath(`channels/${channelId}/${options.type}`);
 
-	if (typeof parameters.continuation !== 'undefined')
-		url.searchParams.set('continuation', parameters.continuation);
+	if (typeof options.continuation !== 'undefined')
+		url.searchParams.set('continuation', options.continuation);
 
-	if (typeof parameters.sortBy !== 'undefined') url.searchParams.set('sort_by', parameters.sortBy);
+	if (typeof options.sortBy !== 'undefined') url.searchParams.set('sort_by', options.sortBy);
+
+	if (isYTBackend() || useEngineFallback('ChannelContent')) {
+		return await getChannelContentYTjs(channelId, options);
+	}
 
 	const resp = await fetchErrorHandle(await fetch(url.toString(), fetchOptions));
 	return await resp.json();
@@ -171,7 +193,14 @@ export async function searchChannelContent(
 	channelId: string,
 	search: string,
 	fetchOptions?: RequestInit
-) {
+): Promise<ChannelContent> {
+	// Not Implemented in YTjs
+	if (isYTBackend()) {
+		return {
+			videos: []
+		};
+	}
+
 	const path = buildPath(`channel/${channelId}/search`);
 	path.search = new URLSearchParams({ q: search }).toString();
 	const resp = await fetchErrorHandle(await fetch(path, fetchOptions));
@@ -182,6 +211,10 @@ export async function getSearchSuggestions(
 	search: string,
 	fetchOptions?: RequestInit
 ): Promise<SearchSuggestion> {
+	if (isYTBackend() || useEngineFallback('SearchSuggestions')) {
+		return getSearchSuggestionsYTjs(search);
+	}
+
 	const path = buildPath('search/suggestions');
 	path.search = new URLSearchParams({ q: search }).toString();
 	const resp = await fetchErrorHandle(await fetch(path, fetchOptions));
@@ -189,34 +222,22 @@ export async function getSearchSuggestions(
 }
 
 export async function getHashtag(tag: string, page: number = 0): Promise<{ results: Video[] }> {
+	// TODO: Implement in YTjs
+	if (isYTBackend()) return { results: [] };
+
 	const resp = await fetchErrorHandle(await fetch(buildPath(`hashtag/${tag}?page=${page}`)));
 	return await resp.json();
-}
-
-export interface SearchOptions {
-	sort_by?: 'relevance' | 'rating' | 'upload_date' | 'view_count';
-	type?: 'video' | 'playlist' | 'channel' | 'all';
-	duration?: 'short' | 'medium' | 'long';
-	date?: 'hour' | 'today' | 'week' | 'month' | 'year';
-	features?: string;
-	page?: string;
 }
 
 export async function getSearch(
 	search: string,
 	options: SearchOptions,
 	fetchOptions?: RequestInit
-): Promise<(Channel | Video | Playlist | HashTag)[]> {
-	if (typeof options.sort_by === 'undefined') {
-		options.sort_by = 'relevance';
-	}
+): Promise<SearchResults> {
+	searchSetDefaults(options);
 
-	if (typeof options.type === 'undefined') {
-		options.type = 'all';
-	}
-
-	if (typeof options.page === 'undefined') {
-		options.page = '1';
+	if (isYTBackend() || useEngineFallback('Search')) {
+		return await getSearchYTjs(search, options);
 	}
 
 	const path = buildPath('search');
@@ -230,6 +251,10 @@ export async function getFeed(
 	page: number,
 	fetchOptions: RequestInit = {}
 ): Promise<Feed> {
+	if (isYTBackend()) {
+		return getFeedYTjs(maxResults, page);
+	}
+
 	const path = buildPath('auth/feed');
 	path.search = new URLSearchParams({
 		max_results: maxResults.toString(),
@@ -241,7 +266,18 @@ export async function getFeed(
 	return await resp.json();
 }
 
+export async function notificationsMarkAsRead(fetchOptions: RequestInit = {}) {
+	// Not support functionality of YTjs
+	if (isYTBackend()) return;
+
+	const path = buildPath('auth/notifications');
+	await fetchErrorHandle(await fetch(path, { ...buildAuthHeaders(), ...fetchOptions }));
+}
+
 export async function getSubscriptions(fetchOptions: RequestInit = {}): Promise<Subscription[]> {
+	if (isYTBackend()) {
+		return getSubscriptionsYTjs();
+	}
 	const resp = await fetchErrorHandle(
 		await fetch(buildPath('auth/subscriptions'), { ...buildAuthHeaders(), ...fetchOptions })
 	);
@@ -252,6 +288,10 @@ export async function amSubscribed(
 	authorId: string,
 	fetchOptions: RequestInit = {}
 ): Promise<boolean> {
+	if (isYTBackend()) {
+		return amSubscribedYTjs(authorId);
+	}
+
 	if (!get(authStore)) return false;
 
 	try {
@@ -265,6 +305,10 @@ export async function amSubscribed(
 }
 
 export async function postSubscribe(authorId: string, fetchOptions: RequestInit = {}) {
+	if (isYTBackend()) {
+		return postSubscribeYTjs(authorId);
+	}
+
 	await fetchErrorHandle(
 		await fetch(buildPath(`auth/subscriptions/${authorId}`), {
 			method: 'POST',
@@ -275,6 +319,10 @@ export async function postSubscribe(authorId: string, fetchOptions: RequestInit 
 }
 
 export async function deleteUnsubscribe(authorId: string, fetchOptions: RequestInit = {}) {
+	if (isYTBackend()) {
+		return deleteUnsubscribeYTjs(authorId);
+	}
+
 	await fetchErrorHandle(
 		await fetch(buildPath(`auth/subscriptions/${authorId}`), {
 			method: 'DELETE',
@@ -289,6 +337,11 @@ export async function getHistory(
 	maxResults: number = 20,
 	fetchOptions: RequestInit = {}
 ): Promise<string[]> {
+	// Not supported functionality of YTjs.
+	if (isYTBackend()) {
+		return [];
+	}
+
 	const resp = await fetchErrorHandle(
 		await fetch(buildPath(`auth/history?page=${page}&max_results=${maxResults}`), {
 			...buildAuthHeaders(),
@@ -302,6 +355,8 @@ export async function deleteHistory(
 	videoId: string | undefined = undefined,
 	fetchOptions: RequestInit = {}
 ) {
+	if (isYTBackend()) return;
+
 	let url = '/api/v1/auth/history';
 	if (typeof videoId !== 'undefined') {
 		url += `/${videoId}`;
@@ -317,6 +372,8 @@ export async function deleteHistory(
 }
 
 export async function postHistory(videoId: string, fetchOptions: RequestInit = {}) {
+	if (isYTBackend()) return;
+
 	await fetchErrorHandle(
 		await fetch(buildPath(`auth/history/${videoId}`), {
 			method: 'POST',
@@ -331,6 +388,10 @@ export async function getPlaylist(
 	page: number = 1,
 	fetchOptions: RequestInit = {}
 ): Promise<PlaylistPage> {
+	if (isYTBackend() || useEngineFallback('Playlist')) {
+		return await getPlaylistYTjs(playlistId);
+	}
+
 	let resp;
 
 	if (get(authStore)) {
@@ -348,6 +409,8 @@ export async function getPlaylist(
 export async function getPersonalPlaylists(
 	fetchOptions: RequestInit = {}
 ): Promise<PlaylistPage[]> {
+	if (isYTBackend()) return [];
+
 	const resp = await fetchErrorHandle(
 		await fetch(buildPath('auth/playlists'), { ...buildAuthHeaders(), ...fetchOptions })
 	);
@@ -355,6 +418,8 @@ export async function getPersonalPlaylists(
 }
 
 export async function deletePersonalPlaylist(playlistId: string) {
+	if (isYTBackend()) return;
+
 	await fetchErrorHandle(
 		await fetch(buildPath(`auth/playlists/${playlistId}`), {
 			method: 'DELETE',
@@ -368,6 +433,8 @@ export async function postPersonalPlaylist(
 	privacy: 'public' | 'private' | 'unlisted',
 	fetchOptions: RequestInit = {}
 ) {
+	if (isYTBackend()) return;
+
 	const headers: Record<string, Record<string, string>> = buildAuthHeaders();
 	headers['headers']['Content-type'] = 'application/json';
 
@@ -389,6 +456,8 @@ export async function addPlaylistVideo(
 	videoId: string,
 	fetchOptions: RequestInit = {}
 ) {
+	if (isYTBackend()) return;
+
 	const headers: Record<string, Record<string, string>> = buildAuthHeaders();
 	headers['headers']['Content-type'] = 'application/json';
 
@@ -409,6 +478,8 @@ export async function removePlaylistVideo(
 	indexId: string,
 	fetchOptions: RequestInit = {}
 ) {
+	if (isYTBackend()) return;
+
 	await fetchErrorHandle(
 		await fetch(buildPath(`auth/playlists/${playlistId}/videos/${indexId}`), {
 			method: 'DELETE',
